+++ index.html
@@ -6,7 +6,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap" rel="stylesheet">
     <script src="https://kit.fontawesome.com/4c1b10f580.js" crossorigin="anonymous"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="shortcut icon" type="image/x-icon" href="public/rocket_favicon1.png">
+    <link rel="shortcut icon" type="image/x-icon" href="/rocket_favicon1.png">
 
 
     <title>EGMA</title>
