+++ src/components/HowThingsWork.jsx
@@ -1,14 +1,6 @@
 import { useRef, useEffect, useState } from 'react';
 import propTypes from 'prop-types';
-
-const steps = [
-  { id: 1, text: "Step 1: Understanding the basics" },
-  { id: 2, text: "Step 2: Setting up the environment" },
-  { id: 3, text: "Step 3: Writing your first code" },
-  { id: 4, text: "Step 4: Testing and debugging" },
-  { id: 5, text: "Step 5: Deployment" }
-];
-
+import { steps } from './constants';
 const Step = ({ step, activeStep, index, progress }) => {
   return (
     <div className={`flex ${index % 2 === 0 ? 'flex-row-reverse' : ''} dark:text-white`}>
