+++ src/components/constants.js
@@ -1,0 +1,25 @@
+export const projects = [
+    { name: "Project 1", image: "bw.webp" },
+    { name: "Project 2", image: "bw.webp" },
+    { name: "Project 3", image: "bw.webp" },
+    { name: "Project 3", image: "bw.webp" },
+    { name: "Project 3", image: "bw.webp" },
+    { name: "Project 3", image: "bw.webp" },
+    { name: "Project 3", image: "bw.webp" },
+];
+
+export const links = [
+    { name: "Services", href: "#" },
+    { name: "Process", href: "#" },
+    { name: "Testimonials", href: "#" },
+    { name: "Work", href: "#" },
+    { name: "Contact", href: "#" },
+];
+
+export const steps = [
+  { id: 1, text: "Step 1: Understanding the basics" },
+  { id: 2, text: "Step 2: Setting up the environment" },
+  { id: 3, text: "Step 3: Writing your first code" },
+  { id: 4, text: "Step 4: Testing and debugging" },
+  { id: 5, text: "Step 5: Deployment" }
+];
